requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
requests==2.31.0
python-decouple==3.8
drf-yasg==1.21.7
Pillow==10.1.0
django-filter==23.3
celery==5.3.4
django-celery-beat==2.5.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
